+++ src/commands/geoLocation.ts
@@ -1,6 +1,8 @@
 import fetch from 'node-fetch';
 import chalk from 'chalk';
 
+import colorifyHeader from '../helpers/colorifyHeader';
+
 const url = 'http://free.ipwhois.io/json/';
 
 const getLocationFromIP = async (url: string) => {
@@ -16,13 +18,13 @@ export default async () => {
     const geoLocation = await getLocationFromIP(url);
 
     console.log(
-      `${chalk.cyan('coordinates')}: (${geoLocation.latitude}, ${
-        geoLocation.longitude
-      })\n` +
+      `${colorifyHeader('Computer geolocation:')}\n` +
+        `${chalk.cyan('coordinates')}: ` +
+        `(${geoLocation.latitude}, ${geoLocation.longitude})\n` +
         `${chalk.cyan('city')}: ${geoLocation.city}\n` +
         `${chalk.cyan('region')}: ${geoLocation.region}\n` +
         `${chalk.cyan('country')}: ${geoLocation.country}\n` +
-        `${chalk.cyan('continent')}: ${geoLocation.continent}\n`
+        `${chalk.cyan('continent')}: ${geoLocation.continent}`
     );
   } catch (err) {
     console.error(chalk.red(`${err}`));