+++ src/helpers/colorifyHeader.ts
@@ -1,0 +1,3 @@
+import chalk = require('chalk');
+
+export default (text: string) => chalk.bold.visible.bgBlueBright(text);