+++ src/index.ts
@@ -4,51 +4,41 @@ import path from 'path';
 import program from 'commander';
 import chalk from 'chalk';
 
-import geoLocationCommand from './commands/geoLocation';
-import diskLocationCommand from './commands/diskLocation';
+import geoLocation from './commands/geoLocation';
+import diskLocation from './commands/diskLocation';
 
 const pkg = require(path.join(__dirname, '../package.json'));
 
-program.version(pkg.version).description(chalk.blue('where is my node'));
-
 program
-  .command('location')
-  .alias('lc')
-  .description('checks node location')
-  .option('-g --geo', 'output computer geolocation')
-  .option('-d --disk', 'output directory on disk')
+  .usage('[options]')
+  .option('-g, --geo', 'output computer geolocation')
+  .option('-d, --disk', 'output directory on disk')
+  .version(pkg.version)
+  .description(chalk.blue('checks node location'))
   .action(async cmd => {
-    if (cmd.geo || cmd.disk) {
-      console.log(`${chalk.magenta('Your node.js is here:')}`);
-    }
-
     if (cmd.geo) {
-      await geoLocationCommand();
+      await geoLocation();
     }
 
     if (cmd.geo && cmd.disk) {
-      console.log(`${chalk.magenta('And here:')}`);
+      console.log('');
     }
 
     if (cmd.disk) {
-      await diskLocationCommand();
+      diskLocation();
     }
   });
 
-program.on('command:*', () => {
-  console.error(chalk.red('Invalid command'));
+program.on('command:*', (commands?: string[]) => {
+  if (commands) {
+    console.error('error: unknown command');
 
-  process.exit(1);
+    process.exit(1);
+  }
 });
 
 if (!process.argv.slice(2).length) {
-  console.warn('No command specified');
-
-  process.exit(1);
-}
-
-if (!process.argv.slice(3).length) {
-  program.command('location').outputHelp();
+  program.outputHelp();
 
   process.exit(1);
 }